+++ src/main/java/game_from_scratch/engine/components/Component.java
@@ -1,13 +1,13 @@
 package game_from_scratch.engine.components;
 
 public abstract class Component {
-    protected long entityId;
+    protected int entityId;
 
-    public long getEntityId() {
+    public int getEntityId() {
         return entityId;
     }
 
-    public void setEntityId(long entityId) {
+    public void setEntityId(int entityId) {
         this.entityId = entityId;
     }
 }
